testing

+++ TidalPy/RadialSolver/starting/common.pyx
@@ -41,7 +41,7 @@ cdef double complex cf_z_calc(
     cdef double complex x, z
     cdef double l2_3, l2_5, l2_7, l2_9, l2_11
     cdef double l2 = <double>degree_l * 2.0
-    cdef Py_ssize_t degree_l_ssizet = <Py_ssize_t>degree_l
+    cdef long degree_l_ssizet = <long>degree_l
 
     if cf_cabs(x_squared) > 0.1:
         # Use real function
